refactor: remove unused `load_and_decode_object_file` function
The main goal of this change is to simplify the implementation. The new
interface forces loading and decoding to happen in two separate steps,
which coincidentally is how this library is used in all our apps. It
makes code more readable, addresses open TODOs, and makes clippy happy.
Build bots should finally turn green now.