fix: instructions are a subset of the code segment
The executable segment in an ELF file (according to the program header)
can include more data than the pure instruction sequence (i.e. what is
covered by the  `.text` section). This change adds a field to `Program`
specifying this range in terms of virtual memory addresses. When such a
`Program` is decoded into a `DecodedProgram` we limit the code segment
to only this range.